Don't generate datasets uselessly
