+++ .github/workflows/ci-workflow.yml
@@ -80,17 +80,26 @@ jobs:
       - name: Generate datasets
         run: bash scripts/generate_binary_demo.sh
 
+      - name: Upload binary-demo
+        uses: actions/upload-artifact@v4
+        with:
+          name: binary-demo
+          path: dataset/binary-demo
+
       - name: Generate and upload tinysnb
         run: |
-          bash scripts/generate_binary_tinysnb.sh
-          s3cmd get s3://kuzu-test/tinysnb/version.txt || echo "0" > version.txt
-          s3cmd --access_key=${GCS_ACCESS_KEY_ID} --secret_key=${GCS_SECRET_ACCESS_KEY} --host=storage.googleapis.com --host-bucket="%(bucket)s.storage.googleapis.com" get s3://kuzudb-test/tinysnb/version.txt || echo "0" > version_gcs.txt
-          if [ "$(cat tinysnb/version.txt)" == "$(cat version.txt)" ] && [ "$(cat tinysnb/version.txt)" == "$(cat version_gcs.txt)" ]; then
+          version_current="$(python3 benchmark/version.py)"
+          echo "0" > version_s3.txt
+          echo "0" > version_gcs.txt
+          s3cmd get --force s3://kuzu-test/tinysnb/version.txt version_s3.txt || true
+          s3cmd --access_key=${GCS_ACCESS_KEY_ID} --secret_key=${GCS_SECRET_ACCESS_KEY} --host=storage.googleapis.com --host-bucket="%(bucket)s.storage.googleapis.com" get --force s3://kuzudb-test/tinysnb/version.txt version_gcs.txt || true
+          if [[ "$version_current" == "$(cat version_s3.txt)" ]] && [[ "$version_current" == "$(cat version_gcs.txt)" ]]; then
             echo "TinySNB dataset is up to date, skipping upload"
-            rm -rf tinysnb version.txt
+            rm -rf tinysnb version_s3.txt version_gcs.txt
             exit 0
           fi
           echo "TinySNB dataset is outdated, uploading..."
+          bash scripts/generate_binary_tinysnb.sh
           s3cmd del -r s3://kuzu-test/tinysnb/
           s3cmd sync ./tinysnb s3://kuzu-test/
           # s3cmd del -r doesn't work on GCS so we individually delete each object in the directory
@@ -99,27 +108,23 @@ jobs:
               s3cmd --access_key=${GCS_ACCESS_KEY_ID} --secret_key=${GCS_SECRET_ACCESS_KEY} --host=storage.googleapis.com --host-bucket="%(bucket)s.storage.googleapis.com" del ${gcs_file}
           done
           s3cmd --access_key=${GCS_ACCESS_KEY_ID} --secret_key=${GCS_SECRET_ACCESS_KEY} --host=storage.googleapis.com --host-bucket="%(bucket)s.storage.googleapis.com" sync ./tinysnb s3://kuzudb-test/
-          rm -rf tinysnb version.txt version_gcs.txt
+          rm -rf tinysnb version_s3.txt version_gcs.txt
 
       - name: Generate and upload ldbc-sf01
         run: |
-          bash scripts/generate_binary_ldbc-sf01.sh
-          s3cmd get s3://kuzu-test/ldbc01/version.txt || echo "0" > version.txt
-          if [ "$(cat ldbc01/version.txt)" == "$(cat version.txt)" ]; then
+          version_current="$(python3 benchmark/version.py)"
+          echo "0" > version_s3.txt
+          s3cmd get --force s3://kuzu-test/ldbc01/version.txt version_s3.txt || true
+          if [[ "$version_current" == "$(cat version_s3.txt)" ]]; then
             echo "LDBC-SF01 dataset is up to date, skipping upload"
-            rm -rf ldbc01 version.txt
+            rm -rf ldbc01 version_s3.txt
             exit 0
           fi
           echo "LDBC-SF01 dataset is outdated, uploading..."
+          bash scripts/generate_binary_ldbc-sf01.sh
           s3cmd del -r s3://kuzu-test/ldbc01/
           s3cmd sync ./ldbc01 s3://kuzu-test/
-          rm -rf ldbc01 version.txt
-
-      - name: Upload binary-demo
-        uses: actions/upload-artifact@v4
-        with:
-          name: binary-demo
-          path: dataset/binary-demo
+          rm -rf ldbc01 version_s3.txt
 
   gcc-build-test:
     name: gcc build & test
